oim/baidu/paddleocr/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "oim-paddleocr"
version = "0.1.0"
description = "Open replacement for paddleocr NIM packaged for uv."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Open NV Ingest Maintainers" }]
dependencies = [
  "fastapi==0.124.4",
  "grpcio==1.67.1",
  "numpy==2.3.5",
  "opencv-contrib-python==4.10.0.84",
  "opentelemetry-api==1.39.1",
  "opentelemetry-exporter-otlp-proto-http==1.39.1",
  "opentelemetry-sdk==1.39.1",
  "oim-common>=0.1.0",
  "paddlepaddle==3.2.1",
  "paddleocr==3.3.2",
  "paddlex[ocr]==3.3.11",
  "pillow==12.0.0",
  "prometheus-client==0.23.1",
  "pydantic-settings==2.12.0",
  "pydantic==2.12.5",
  "requests==2.32.5",
  "tritonclient[http]==2.60.0",
  "tritonclient[grpc]==2.60.0",
  "uvicorn[standard]==0.38.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "ruff>=0.7.0", "mypy>=1.13.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/oim_paddleocr"]

[tool.uv]
managed = true

[tool.uv.sources]
oim-common = { workspace = true }
